refactor: move production code to their own files
