+++ src/infra/database/postgres/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './task'
