+++ src/infra/database/postgres/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './task-repository'
