+++ tests/infra/database/postgres/repositories/task-repository.spec.ts
@@ -1,34 +1,7 @@
 import { Task } from '@/domain/entities'
-import { Column, Entity, getRepository, PrimaryGeneratedColumn } from 'typeorm'
+import { PgTask } from '@/infra/database/postgres/entities'
+import { PgTaskRepository } from '@/infra/database/postgres/repositories'
 import { makeDbConnection } from '@/tests/infra/database/postgres/mocks/connection'
-import { AddTaskRepository } from '@/data/contracts'
-
-@Entity({ name: 'Task' })
-class PgTask {
-  @PrimaryGeneratedColumn()
-  id!: number
-
-  @Column()
-  title!: string
-
-  @Column()
-  description!: string
-
-  @Column({ type: 'boolean' })
-  isFavorite!: boolean
-
-  @Column({ type: 'boolean' })
-  isComplete!: boolean
-}
-
-class PgTaskRepository {
-  async save (task: Task): Promise<AddTaskRepository.Result> {
-    const pgTaskRepository = getRepository(PgTask)
-
-    const { id } = await pgTaskRepository.save(task)
-    return { id: id.toString() }
-  }
-}
 
 const mockTask: Task = {
   title: 'any_title',
